+++ src/main/java/io/nagurea/smsupsdk/accountmanaging/account/retrieve/response/subaccount/RetrieveSubaccountsResponse.java
@@ -1,0 +1,15 @@
+package io.nagurea.smsupsdk.accountmanaging.account.retrieve.response.subaccount;
+
+import io.nagurea.smsupsdk.common.response.APIResponse;
+import lombok.Builder;
+import lombok.ToString;
+
+@ToString
+public class RetrieveSubaccountsResponse extends APIResponse<RetrieveSubaccountsResultResponse> {
+
+    @Builder
+    public RetrieveSubaccountsResponse(String uid, Integer statusCode, String additionalMessage, RetrieveSubaccountsResultResponse effectiveResponse) {
+        super(uid, statusCode, additionalMessage, effectiveResponse);
+    }
+
+}