+++ src/test/java/io/nagurea/smsupsdk/accountmanaging/account/retrieve/RetrieveAccountServiceIntTest.java
@@ -3,6 +3,7 @@ package io.nagurea.smsupsdk.accountmanaging.account.retrieve;
 import io.nagurea.smsupsdk.accountmanaging.account.retrieve.response.account.AccountInfo;
 import io.nagurea.smsupsdk.accountmanaging.account.retrieve.response.account.RetrieveAccountResultResponse;
 import io.nagurea.smsupsdk.accountmanaging.account.retrieve.response.subaccount.RetrieveSubaccountResultResponse;
+import io.nagurea.smsupsdk.accountmanaging.account.retrieve.response.subaccount.RetrieveSubaccountsResultResponse;
 import io.nagurea.smsupsdk.accountmanaging.account.retrieve.response.subaccount.SubaccountInfo;
 import io.nagurea.smsupsdk.common.status.ResponseStatus;
 import org.junit.jupiter.api.AfterAll;
@@ -40,42 +41,42 @@ class RetrieveAccountServiceIntTest {
     @BeforeAll
     public static void startMockSMSUpServer(){
         mockServer = ClientAndServer.startClientAndServer("localhost", 4242, 4242);
-//        mockServer.when(
-//                request()
-//                        .withPath("/account")
-//                        .withMethod("GET")
-//                        .withHeader("Authorization", EXPECTED_TOKEN)
-//        ).respond(
-//                HttpResponse.response()
-//                        .withStatusCode(200)
-//                        .withBody(
-//                            "{\n" +
-//                                    "    \"status\": 1,\n" +
-//                                    "    \"message\": \"OK\",\n" +
-//                                    "    \"account\": {\n" +
-//                                    "      \"client_id\": \"1010101\",\n" +
-//                                    "      \"email\": \"mr@robot.com\",\n" +
-//                                    "      \"firstname\": \"Elliot\",\n" +
-//                                    "      \"lastname\": \"Alderson\",\n" +
-//                                    "      \"city\": \"New York\",\n" +
-//                                    "      \"type\": \"private\",\n" +
-//                                    "      \"company\": \"Allsafe\",\n" +
-//                                    "      \"phone\": \"41781234567\",\n" +
-//                                    "      \"address1\": \"unknown address\",\n" +
-//                                    "      \"address2\": null,\n" +
-//                                    "      \"zip\": \"10040\",\n" +
-//                                    "      \"country\": \"\",\n" +
-//                                    "      \"country_code\": \"US\",\n" +
-//                                    "      \"lang\": \"EN\",\n" +
-//                                    "      \"credits\": \"0\",\n" +
-//                                    "      \"unlimited\": \"1\",\n" +
-//                                    "      \"description\": \"Fsociety\",\n" +
-//                                    "      \"senderid\": \"\",\n" +
-//                                    "      \"status\": \"1\"\n" +
-//                                    "    }\n" +
-//                                    "}"
-//                        )
-//        );
+        mockServer.when(
+                request()
+                        .withPath("/account")
+                        .withMethod("GET")
+                        .withHeader("Authorization", EXPECTED_TOKEN)
+        ).respond(
+                HttpResponse.response()
+                        .withStatusCode(200)
+                        .withBody(
+                            "{\n" +
+                                    "    \"status\": 1,\n" +
+                                    "    \"message\": \"OK\",\n" +
+                                    "    \"account\": {\n" +
+                                    "      \"client_id\": \"1010101\",\n" +
+                                    "      \"email\": \"mr@robot.com\",\n" +
+                                    "      \"firstname\": \"Elliot\",\n" +
+                                    "      \"lastname\": \"Alderson\",\n" +
+                                    "      \"city\": \"New York\",\n" +
+                                    "      \"type\": \"private\",\n" +
+                                    "      \"company\": \"Allsafe\",\n" +
+                                    "      \"phone\": \"41781234567\",\n" +
+                                    "      \"address1\": \"unknown address\",\n" +
+                                    "      \"address2\": null,\n" +
+                                    "      \"zip\": \"10040\",\n" +
+                                    "      \"country\": \"\",\n" +
+                                    "      \"country_code\": \"US\",\n" +
+                                    "      \"lang\": \"EN\",\n" +
+                                    "      \"credits\": \"0\",\n" +
+                                    "      \"unlimited\": \"1\",\n" +
+                                    "      \"description\": \"Fsociety\",\n" +
+                                    "      \"senderid\": \"\",\n" +
+                                    "      \"status\": \"1\"\n" +
+                                    "    }\n" +
+                                    "}"
+                        )
+        );
 
         mockServer.when(
                 request()
@@ -109,6 +110,43 @@ class RetrieveAccountServiceIntTest {
                                         "        \"senderid\": \"\",\n" +
                                         "        \"status\": \"1\"\n" +
                                         "      }\n" +
+                                        "}"
+                        )
+        );
+
+        mockServer.when(
+                request()
+                        .withPath("/sub-accounts")
+                        .withMethod("GET")
+                        .withHeader("Authorization", EXPECTED_TOKEN)
+        ).respond(
+                HttpResponse.response()
+                        .withStatusCode(200)
+                        .withBody(
+                                "{\n" +
+                                        "  \"status\": 1,\n" +
+                                        "  \"message\": \"OK\",\n" +
+                                        "  \"sub-accounts\": [\n" +
+                                        "    {\n" +
+                                        "      \"client_id\": \"1010101\",\n" +
+                                        "      \"email\": \"S@ch.fr\",\n" +
+                                        "      \"firstname\": \"Yoh\",\n" +
+                                        "      \"lastname\": \"Asakura\",\n" +
+                                        "      \"city\": \"Izumo\",\n" +
+                                        "      \"phone\": \"41781234567\",\n" +
+                                        "      \"address1\": \"Somewhere\",\n" +
+                                        "      \"address2\": \"Elsewere\",\n" +
+                                        "      \"zip\": \"36520\",\n" +
+                                        "      \"country\": \"\",\n" +
+                                        "      \"country_code\": \"FR\",\n" +
+                                        "      \"lang\": \"FR\",\n" +
+                                        "      \"credits\": \"0\",\n" +
+                                        "      \"unlimited\": \"1\",\n" +
+                                        "      \"description\": \"Shaman King Company\",\n" +
+                                        "      \"senderid\": \"\",\n" +
+                                        "      \"status\": \"1\"\n" +
+                                        "    }\n" +
+                                        "  ]\n" +
                                         "}"
                         )
         );
@@ -190,6 +228,44 @@ class RetrieveAccountServiceIntTest {
 
         // when
         final RetrieveSubaccountResultResponse actualAccountResponse = retrieveAccountService.retrieveSubaccount(YOUR_TOKEN, SUB_ACCCOUNT_ID).getEffectiveResponse();
+
+        // then
+        assertEquals(ResponseStatus.OK, actualAccountResponse.getStatus());
+        assertEquals("OK", actualAccountResponse.getMessage());
+        assertEquals(expectedGetListsResponse, actualAccountResponse);
+
+    }
+
+    @Test
+    void retrieveSubaccounts() throws IOException {
+        // given
+        RetrieveSubaccountsResultResponse expectedGetListsResponse = RetrieveSubaccountsResultResponse.builder()
+                .message("OK")
+                .responseStatus(ResponseStatus.OK)
+                .subaccount(SubaccountInfo.builder()
+                        .clientId("1010101")
+                        .email("S@ch.fr")
+                        .firstname("Yoh")
+                        .lastname("Asakura")
+                        .city("Izumo")
+                        .phone("41781234567")
+                        .address1("Somewhere")
+                        .address2("Elsewere")
+                        .zip("36520")
+                        .country("")
+                        .countryCode("FR")
+                        .lang("FR")
+                        .credits("0")
+                        .unlimited("1")
+                        .description("Shaman King Company")
+                        .senderid("")
+                        .status("1")
+                        .build())
+                .build();
+
+
+        // when
+        final RetrieveSubaccountsResultResponse actualAccountResponse = retrieveAccountService.retrieveSubaccounts(YOUR_TOKEN).getEffectiveResponse();
 
         // then
         assertEquals(ResponseStatus.OK, actualAccountResponse.getStatus());